+++ pozuelo/scripts/azureCLILogin.sh
@@ -1,0 +1,2 @@
+# https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/guides/azure_cli
+az login