+++ pozuelo/scripts/sshKeyGen.sh
@@ -1,0 +1,4 @@
+# Generación de claves para ejecutar en el Ubuntu container
+
+apt-get install openssh-client
+ssh-keygen -t rsa