+++ pozuelo/scripts/terraformInstall.sh
@@ -1,11 +1,6 @@
-# Se ejecuta une vez dentro del container Ubuntu 20.04.4
-
-apt-get update
-apt-get -y install sudo
-
 # Documentación oficial de TERRAFORM
+# https://www.terraform.io/downloads
 # https://learn.hashicorp.com/tutorials/terraform/install-cli
-
 sudo apt-get update && sudo apt-get install -y gnupg software-properties-common curl
 curl -fsSL https://apt.releases.hashicorp.com/gpg | sudo apt-key add -
 sudo apt-add-repository "deb [arch=amd64] https://apt.releases.hashicorp.com $(lsb_release -cs) main"